+++ index.js
@@ -1,7 +1,6 @@
 const NATS = require('nats')
 const logger = require('./lib/logger')
 const merge = require('lodash/merge')
-const keys = require('lodash/keys')
 
 
 function Wrapper(options) {
@@ -26,11 +25,33 @@ function Wrapper(options) {
         })
     }
 
-    this.subscribe = function (subject, opts, done) {
+    this.respond = function (subject, error, message) {
+        if (error) {
+            logger.debug('sending error response to', subject, error)
+            nats.publish(subject, JSON.stringify({error: error}), function() {
+                logger.debug('error response sent to', subject)
+            })
+        } else {
+            logger.debug('sending response to', subject, message)
+            nats.publish(subject, JSON.stringify(message), function() {
+                logger.debug('response sent to', subject)
+            })
+        }
+    }
+
+    this.subscribe = function (subject, done) {
         logger.debug('subscribing to', subject)
-        nats.subscribe(subject, opts, function(message, reply, subject) {
-            logger.debug('received event via subscription', subject, message)
+        nats.subscribe(subject, function(message, reply, subject) {
+            logger.debug('responding to', subject, message)
             done(JSON.parse(message), reply, subject)
+        })
+    }
+
+    this.process = function (subject, group, done) {
+        logger.debug('subscribing to:', subject, 'in group:', group)
+        nats.subscribe(subject, {queue: group}, function(message, reply, subject) {
+            logger.debug('processing', subject, message)
+            done(JSON.parse(message), subject)
         })
     }
 
@@ -49,10 +70,6 @@ function Wrapper(options) {
                 logger.error('request ended in error:', error.message || error.detail)
                 return done(new Error(error.message || error.detail))
             }
-
-            // if response is wrapped in a single key - unwrap it
-            const keys_ = keys(res)
-            if (keys_.length == 1) res = res[keys_[0]]
 
             logger.debug('<<<', res)
             return done(null, res)