+++ test.js
@@ -16,19 +16,10 @@ const wrapper = Wrapper({connection: connection})
 describe('request', function () {
 
     it('performs a request and returns successful result unwrapped', function (done) {
-        NATSrequestOne.callsArgWith(4, JSON.stringify({error: null, account: {id: 'ACC001'}}))
+        NATSrequestOne.callsArgWith(4, JSON.stringify({id: 'ACC001'}))
         wrapper.request('account.get', {id: 'ACC001'}, function (error, account) {
             assert.isNull(error)
             assert.equal(account.id, 'ACC001')
-            done()
-        })
-    })
-
-    it('performs a request and returns single value if response has single key', function (done) {
-        NATSrequestOne.callsArgWith(4, JSON.stringify({error: null, id: 'ACC001'}))
-        wrapper.request('account.get', {id: 'ACC001'}, function (error, response) {
-            assert.isNull(error)
-            assert.equal(response, 'ACC001')
             done()
         })
     })
@@ -64,11 +55,24 @@ describe('request', function () {
 
 describe('subscribe', function () {
 
-    it('performs a subscription to a subject', function (done) {
-        NATSsubscribe.callsArgWith(2, JSON.stringify({data: {one: 'two'}}), null, 'test.event.happened')
-        wrapper.subscribe('test.event.happened', {}, function (message, replyTo, subject) {
+    it('performs a response-subscription to a subject', function (done) {
+        NATSsubscribe.callsArgWith(1, JSON.stringify({data: {one: 'two'}}), null, 'test.event.happened')
+        wrapper.subscribe('test.event.happened', function (message, replyTo, subject) {
             assert.equal(message.data.one, 'two')
             assert.isNull(replyTo)
+            assert.equal(subject, 'test.event.happened')
+            done()
+        })
+    })
+
+})
+
+describe('process', function () {
+
+    it('performs a process-subscription to a subject', function (done) {
+        NATSsubscribe.callsArgWith(2, JSON.stringify({data: {one: 'two'}}), null, 'test.event.happened')
+        wrapper.process('test.event.happened', 'worker-group', function (message, subject) {
+            assert.equal(message.data.one, 'two')
             assert.equal(subject, 'test.event.happened')
             done()
         })
@@ -81,6 +85,24 @@ describe('publish', function () {
     it('performs publishing to a subject', function (done) {
         NATSpublish.callsArgWith(2, null)
         wrapper.publish('test.event.happened', {message: 'yay!'})
+        assert.isTrue(NATSpublish.called)
+        done()
+    })
+
+})
+
+describe('respond', function () {
+
+    it('responds with an error', function (done) {
+        NATSpublish.callsArgWith(2, null)
+        wrapper.respond('test.request', new Error('something bad happened'))
+        assert.isTrue(NATSpublish.called)
+        done()
+    })
+
+    it('responds normally', function (done) {
+        NATSpublish.callsArgWith(2, null)
+        wrapper.respond('test.request', null, {response: 'ok'})
         assert.isTrue(NATSpublish.called)
         done()
     })