+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+[1.1.0] - 2017-04-15
+--------------------
+- [x] ADDED readme
+- [x] ADDED `respond` method fix #4
+- [x] ADDED `process` method to subscribe to a queue group fix #4
+- [x] REMOVE `options` argument from `subscribe` #4
+
+
 [1.0.0] - 2017-04-12
 --------------------
 - [x] CHANGED request response to return unwrapped object fix #3